+++ src/ecommerce/.editorconfig
@@ -1,0 +1,23 @@
+# top-most EditorConfig file
+root = true
+
+# all files
+[*]
+end_of_line = lf
+insert_final_newline = true
+indent_style = space
+indent_size = 2
+charset = utf-8
+trim_trailing_whitespace = true
+max_line_length = 80
+quote_type = single
+
+[*.{js,ts}]
+quote_type = single
+curly_bracket_next_line = false
+spaces_around_brackets = inside
+indent_brace_style = BSD KNF
+
+# HTML
+[*.html]
+quote_type = double
