+++ src/ecommerce/.eslintignore
@@ -1,0 +1,3 @@
+.cache
+build
+**/node_modules/**
