+++ src/ecommerce/.eslintrc
