+++ src/ecommerce/.github/ISSUE_TEMPLATE/other.yml
@@ -1,0 +1,9 @@
+name: Other
+description: Any other topic you want to talk about.
+body:
+  - type: textarea
+    attributes:
+      label: Other
+      description: What would you want to talk about ?
+    validations:
+      required: false
