+++ src/ecommerce/.github/dependabot.yml
@@ -1,0 +1,17 @@
+
+version: 2
+updates:
+- package-ecosystem: npm
+  directory: "/"
+  schedule:
+    interval: weekly
+    time: "04:00"
+  open-pull-requests-limit: 10
+  labels:
+  - skip-changelog
+  - dependencies
+  versioning-strategy: increase
+  rebase-strategy: disabled
+  ignore:
+      - dependency-name: "eslint*"
+      - dependency-name: "*prettier*"
