+++ src/ecommerce/bors.toml
@@ -1,0 +1,5 @@
+status = [
+  'linter_check'
+]
+# 1 hour timeout
+timeout-sec = 3600
