+++ src/ecommerce/components/layout/Blur.jsx
