+++ src/ecommerce/components/layout/NavBar.jsx
