+++ src/ecommerce/next.config.js
@@ -1,6 +1,6 @@
 module.exports = {
   reactStrictMode: true,
   images: {
-    domains: ['']
-  }
+    domains: [''],
+  },
 }
