+++ src/ecommerce/package.json
@@ -23,6 +23,8 @@
   "devDependencies": {
     "eslint": "7.29.0",
     "eslint-config-next": "11.0.0",
-    "prettier": "^2.6.2"
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "prettier": "^2.3.2"
   }
 }
